+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Adds a new set of APIs for defining and CloudWatch metrics and creating alarms from them. See
   [aws.cloudwatch.Metric] for more details, and see [aws.lambda.metrics.duration] as an example of
   a newly exposed easy-to-use metric.
+- Dashboards can easily be created from the above [aws.cloudwatch.Metric] objects using the new
+  [aws.cloudwatch.Dashboard.fromBody] helper.  See [here](https://github.com/pulumi/pulumi-aws/blob/master/examples/dashboards/index.ts) for an example.
 
 ## 0.18.1 (Released April 5th, 2019)